feat: add general dto use case
