+++ internal/events/usecase/list_events.go
@@ -6,19 +6,6 @@ type ListEventsOutputDTO struct {
 	Events []EventDTO
 }
 
-type EventDTO struct {
-	ID           string  `json:"id"`
-	Name         string  `json:"name"`
-	Location     string  `json:"location"`
-	Organization string  `json:"organization"`
-	Rating       string  `json:"rating"`
-	Date         string  `json:"date"`
-	Capacity     int     `json:"capacity"`
-	Price        float64 `json:"price"`
-	PartnerID    int     `json:"partner_id"`
-	ImageURL     string  `json:"image_url"`
-}
-
 type ListEventsUseCase struct {
 	repo domain.EventRepository
 }
